pyproject.toml
[tool.poetry]
name = "raspberrify"
version = "0.1.0"
description = ""
authors = ["johanneshardt <46093733+johanneshardt@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Pillow = "^8.2.0"
requests = "^2.25.1"
spotipy = "^2.18.0"
sense-hat = "^2.2.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = {version = "^21.5b0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
